requirements.txt
yfinance==0.2.36
streamlit==1.32.0
openai==1.12.0
chromadb==0.4.22
langchain==0.1.9
langchain-openai==0.0.8
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
